Update Module
I recently found out that "gg.searchPointer" incorrectly searches for links on some devices (I think this is due to the fact that gg incorrectly reads memory on these devices). So I wrote my "searchPointer", but it works slowly but correctly. Because of this, the search for objects has slowed down. In the future I will try to make this search faster.